+++ tests/testthat/test-profileLikelihood.R
@@ -49,3 +49,29 @@ start, length, event, x1, x2
     expect_equivalent(coef(fit)["x1"], 0)
 })
 
+test_that("Check adapative profiling likelihood", {
+    test <- read.table(header=T, sep = ",", text = "
+start, length, event, x1, x2
+0, 4,  1,0,0
+0, 3.5,1,2,0
+0, 3,  0,0,1
+0, 2.5,1,0,1
+0, 2,  1,1,1
+0, 1.5,0,1,0
+0, 1,  1,1,0
+")
+    data <- createCyclopsData(Surv(length, event) ~ x1 + x2, data = test,
+                              modelType = "cox")
+    fit <- fitCyclopsModel(data)
+
+    expect_error(getCyclopsProfileLogLikelihood(fit, "x1"))
+    expect_error(getCyclopsProfileLogLikelihood(fit, "x1", x = c(0,1), bounds = c(0.8, 0.9)))
+    expect_error(getCyclopsProfileLogLikelihood(fit, "x1", bounds = c(0.8, 0.8)))
+
+    out <- getCyclopsProfileLogLikelihood(fit, "x1", bounds = c(0, 2), initialGridSize = 10)
+    expect_gt(nrow(out), 10)
+
+    argMax <- out$point[which(out$value == max(out$value))]
+    expect_equal(coef(fit)["x1"], argMax)
+})
+